Fixed casting problem with authorize
git-svn-id: http://127.0.0.1/svn/trunk@393 02e793a8-47cb-4219-9cc1-7115cfedab66

+++ src/starcorp/server/entitystore/HibernateStore.java
@@ -560,7 +560,7 @@ public class HibernateStore implements IEntityStore {
 		Map<String, Object> map = prepareParameters("email", email);
 		prepareParameters(map, "password", password);
 		beginTransaction();
-		int count = (Integer) createQuery(q, map).uniqueResult();
+		long count = (Long) createQuery(q, map).uniqueResult();
 		commit();
 		return count > 0;
 	}
